+++ package.json
@@ -14,8 +14,8 @@
     "terser": "for p in ./lib/{index,html}.js; do terser --compress --mangle toplevel --output $p -- $p; done",
     "clean": "rm -rf ./{lib,docs}",
     "test": "jest",
-    "perf": "node --max-old-space-size=4096 src/test/perf.js",
-    "docs": "typedoc --mode library --name 'TagSoup' --includeVersion --out ./docs ./src/main/html.ts",
+    "perf": "node --max-old-space-size=4096 ./src/test/perf.js",
+    "docs": "typedoc ./src/main/html.ts",
     "publish-docs": "set -ex; [ -d ./docs ]; [[ ! $(git status --porcelain) ]]; branch=$(git rev-parse --abbrev-ref HEAD); sha=$(git rev-parse --short HEAD); t=$(mktemp -d); cp -R ./docs/ $t; git checkout ghpages; ls -A | grep -wv .git | xargs rm -rf; cp -R $t/ .; git add .; git commit -m \"Updated docs ($sha)\"; git push; git checkout $branch",
     "release-docs": "set -e; npm run clean; npm run docs; npm run publish-docs"
   },
