+++ src/main/html.ts
@@ -1,2 +1,4 @@
 export * from './xml';
+export * from './createHtmlSaxParser';
 export * from './createHtmlDomParser';
+export * from './createFromHtmlCharName';
