+++ src/main/xml.ts
@@ -1,16 +1,5 @@
-export {
-  Attribute,
-  StartTagCallback,
-  EndTagCallback,
-  DataCallback,
-  SaxParserDialectOptions,
-  SaxParserCallbacks,
-  SaxParserOptions,
-  SaxParser,
-  createSaxParser,
-} from './createSaxParser';
+export * from './createSaxParser';
 export * from './createForgivingSaxParser';
 export * from './createDomParser';
 export * from './createXmlDomParser';
 export * from './createEntitiesDecoder';
-export * from './fromXmlCharName';
