+++ src/test/createDomParser.test.ts
@@ -24,13 +24,13 @@ describe('createDomParser', () => {
   });
 
   it('parses text', () => {
-    expect(parser.commit('okay')).toEqual([
+    expect(parser.parse('okay')).toEqual([
       {data: 'okay', start: 0, end: 4},
     ]);
   });
 
   it('parses tag with text', () => {
-    expect(parser.commit('<a>okay</a>')).toEqual([
+    expect(parser.parse('<a>okay</a>')).toEqual([
       {
         tagName: 'a', start: 0, end: 11, attrs: {length: 0}, children: [
           {data: 'okay', start: 3, end: 7},
@@ -40,7 +40,7 @@ describe('createDomParser', () => {
   });
 
   it('parses attributes', () => {
-    expect(parser.commit('<a foo=bar></a>')).toEqual([
+    expect(parser.parse('<a foo=bar></a>')).toEqual([
       {tagName: 'a', start: 0, end: 15, attrs: {length: 1, 0: {name: 'foo', value: 'bar', start: 3, end: 10}}, children: []},
     ]);
   });
@@ -51,7 +51,7 @@ describe('createDomParser', () => {
       isVoidContent: (tagName) => tagName === 'a',
     });
 
-    expect(parser.commit('<a><a>')).toEqual([
+    expect(parser.parse('<a><a>')).toEqual([
       {tagName: 'a', start: 0, end: 3, attrs: {length: 0}, children: []},
       {tagName: 'a', start: 3, end: 6, attrs: {length: 0}, children: []},
     ]);
@@ -63,14 +63,14 @@ describe('createDomParser', () => {
       isVoidContent: (tagName) => tagName === 'a',
     });
 
-    expect(parser.commit('<a><b></b></a>')).toEqual([
+    expect(parser.parse('<a><b></b></a>')).toEqual([
       {tagName: 'a', start: 0, end: 3, attrs: {length: 0}, children: []},
       {tagName: 'b', start: 3, end: 10, attrs: {length: 0}, children: []},
     ]);
   });
 
   it('parses nested tags', () => {
-    expect(parser.commit('<a><b></b></a>')).toEqual([
+    expect(parser.parse('<a><b></b></a>')).toEqual([
       {
         tagName: 'a', start: 0, end: 14, attrs: {length: 0}, children: [
           {tagName: 'b', start: 3, end: 10, attrs: {length: 0}, children: []},
@@ -80,7 +80,7 @@ describe('createDomParser', () => {
   });
 
   it('parses nested tags that are closed in wrong order', () => {
-    expect(parser.commit('<a><b></a></b>')).toEqual([
+    expect(parser.parse('<a><b></a></b>')).toEqual([
       {
         tagName: 'a', start: 0, end: 10, attrs: {length: 0}, children: [
           {tagName: 'b', start: 3, end: 6, attrs: {length: 0}, children: []},
@@ -90,7 +90,7 @@ describe('createDomParser', () => {
   });
 
   it('parses nested tags that are closed in wrong order with matching parent', () => {
-    expect(parser.commit('<b><a><b></a></b>')).toEqual([
+    expect(parser.parse('<b><a><b></a></b>')).toEqual([
       {
         tagName: 'b', start: 0, end: 17, attrs: {length: 0}, children: [
           {
@@ -106,7 +106,7 @@ describe('createDomParser', () => {
   it('ignores unmatched closing tags', () => {
     parser = createDomParser(domParserOptions);
 
-    expect(parser.commit('<a></b>eee')).toEqual([
+    expect(parser.parse('<a></b>eee')).toEqual([
       {
         tagName: 'a', start: 0, end: 10, attrs: {length: 0}, children: [
           {data: 'eee', start: 7, end: 10},
