+++ typedoc.json
@@ -1,0 +1,42 @@
+{
+  "name": "TagSoup",
+  "mode": "library",
+  "includeVersion": true,
+  "out": "./docs",
+  "toc": [
+    "DomParserOptions",
+    "DomParser",
+    "createDomParser",
+    "createXmlDomParser",
+    "createHtmlDomParser",
+
+    "ForgivingSaxParserOptions",
+    "createForgivingSaxParser",
+
+    "EntitiesDecoderOptions",
+    "createEntitiesDecoder",
+
+    "FromCharCodeOptions",
+    "createFromCharCode",
+
+    "FromHtmlCharNameOptions",
+    "createFromHtmlCharName",
+
+    "HtmlDomParserOptions",
+    "createHtmlDomParser",
+
+    "HtmlSaxParserOptions",
+    "createHtmlSaxParser",
+
+    "DomNodeType",
+    "DomNode",
+    "DomAttrMap",
+    "DomElement",
+    "DomText",
+    "createXmlDomParser",
+
+    "SaxParserOptions",
+    "SaxParser",
+    "createSaxParser"
+  ]
+}
